feat: consolidate pending feat/graph-3d work + session fixes
This is a large consolidation commit (106 files, +12,054/-900) that bundles
two logical layers of work that accumulated on the feat/graph-3d branch:

## Prior-session work (clean-architecture + feature-dev)

- Repositories layer: backend/app/repositories/ — simulation_repo, project_repo,
  memory_repo, protocols, simulation_persistence split out of engine code so
  the orchestrator no longer owns its own DB sessions
- Services layer: backend/app/services/ — simulation_service, community_opinion_service,
  notification_service, ports — clean separation between HTTP handlers and
  orchestrator/engine, with session lifecycle managed at the service boundary
- Conversation threads: thread_capture pipeline, ThreadMessageRow ORM,
  22_CONVERSATION_THREAD_SPEC tests (test_22_conversation_threads.py)
- Expert LLM engine: richer expert evaluation with SLM fallback
  (23_EXPERT_LLM_SPEC tests in test_23_expert_llm.py)
- Community opinion feature: community_opinion model, service, API,
  CommunityOpinionPanel and pages, migration e1_community_opinion
- LLM cache observability: test_24_cache_observability covering the vcache hit
  path + tier distribution
- Frontend component split-outs: DecidePanel, EmergentEventsPanel,
  EliteLLMNarrativePanel, OverallOpinionPanel, FormProgressBanner,
  SimilarityWarningBanner, WorkflowStepper, GraphLegend, ZoomTierBadge —
  all extracted from their parent containers with matching test files
- ArchitectureInvariants.test.ts + communitySimilarity.test.ts structural guards
- test_21_simulation_quality split into p1/p2 files + test_21_memory_pgvector
- test_25_simulation_service + test_26_community_opinion service-layer tests
- Bigint random_seed migration (d1_bigint_random_seed) — seed column widened
  so values outside int32 range don't overflow

## This session's fixes (documented in CHANGELOG Unreleased)

- **Propagation animation restored** — GraphPanel's active-link keys were
  built from agent UUIDs while linkDirectionalParticles looked them up by
  graph node_ids, so particles never drew. Translation now lives in
  propagationAnimationUtils.ts (buildAgentIdToNodeId + buildActivePropLinks)
  and is exercised by the same regression tests the component uses
- **Low-centrality agents propagate again** — the agent tick path used
  InfluenceLayer which missed the Round 7-d floor + sigmoid emotion
  smoothing that propagation_model.py already had. Both paths now call
  the shared propagation_calibration.propagation_probability() so future
  calibration tweaks live in one file
- **Startup deadlock on /api/v1/projects/** — lifespan split into two
  short-lived transactions with SET LOCAL lock_timeout = '10s', and
  metadata.create_all is skipped entirely when alembic_version is present
- **community_name in agent responses** — AgentDetailResponse gains a
  community_name field, resolved via a cached community_uuid → cc.name
  map (was O(N) graph walk per inspector click, now O(1))
- **Dynamic community palette in 3D graph** — palette is derived from
  the live graph instead of the hardcoded A/B/C/D/E default, so real sims
  with "mainstream"/"skeptics"/etc get real colors. Fallback color is
  hashed from the community id for stability across re-fetches
- **Graph node labels** use the graph node_id (Agent #42) instead of the
  first-8-chars of the agent UUID, which were identical for every
  deterministic-seed agent
- **Graph overlay layout** — 3D Controls hint moved to bottom-right,
  community legend raised 200px, GraphLegend gained a bottomOffsetPx prop
  so the stacking stays coherent
- **Regression tests** — test_01_influence gains
  test_round_7d_low_influence_agents_still_propagate and
  test_round_7d_negative_emotion_factor_still_propagates; PropagationAnimation
  test suite gains 5 tests that exercise the real utility functions

## Verification at commit time

- backend: test_01_influence (10/10), test_07_propagation_pairs (29/29),
  test_04_community_orchestrator, and the agent/influence/propagation/project
  filter (205/205) all green earlier in the session
- frontend: 656/656 pass (40 files) — fresh run at commit time
- tsc -b: clean on all touched files (pre-existing baseline errors in
  communitySimilarity.test.ts and DecidePanel.tsx are not mine)
- Live end-to-end: fresh sim produces non-empty propagation_pairs with
  agent UUIDs that resolve to valid node_ids in the network graph;
  /api/v1/simulations/{id}/agents/{id} returns community_name: "Alpha"

+++ backend/app/models/simulation.py
@@ -3,7 +3,7 @@ SPEC: docs/spec/08_DB_SPEC.md#simulations
 """
 import uuid
 from datetime import datetime
-from sqlalchemy import String, Integer, Float, Text, DateTime, ForeignKey, Index, UniqueConstraint, func
+from sqlalchemy import String, Integer, BigInteger, Float, Text, DateTime, ForeignKey, Index, UniqueConstraint, func
 from sqlalchemy.dialects.postgresql import UUID, JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -24,7 +24,9 @@ class Simulation(Base):
     max_steps: Mapped[int] = mapped_column(Integer, nullable=False, default=50)
     config: Mapped[dict] = mapped_column(JSONB, nullable=False)
     network_metrics: Mapped[dict | None] = mapped_column(JSONB)
-    random_seed: Mapped[int | None] = mapped_column(Integer)
+    # BigInteger so seeds generated via ``os.urandom(4)`` (unsigned 32-bit,
+    # up to 4,294,967,295) fit safely — signed INTEGER tops out at ~2.1B.
+    random_seed: Mapped[int | None] = mapped_column(BigInteger)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     started_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
     completed_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True))
